# server/openapi.yaml
openapi: 3.0.3
info:
  title: OpenTriologue API
  version: 1.0.0
  description: |
    OpenAPI contract for BYOA and project-scoped collaboration flows.
    Focus in this first version: health, BYOA message ingest, project access,
    team invite, and task create/update.
servers:
  - url: http://localhost:4001
    description: Local API
  - url: https://opentriologue.ai
    description: Production API
...
paths:
...
  /api/secrets:
    get:
      tags: [Secrets]
      summary: List user secrets (cursor-paginated)
      operationId: listUserSecrets
      security:
        - BearerJwt: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: Secret ID cursor from previous pageInfo.nextCursor
        - name: scope
          in: query
          required: false
          schema:
            type: string
          description: all | unlinked | {projectId}
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive search in name, description, project name, lastUsedBy
        - name: legacy
          in: query
          required: false
          schema:
            type: boolean
          description: If true, returns legacy array response instead of paginated object
      responses:
        "200":
          description: Paginated secrets (or legacy array when legacy=true)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SecretListResponse"
                  - $ref: "#/components/schemas/SecretListLegacyResponse"
        "400":
          description: Invalid cursor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Secrets]
      summary: Create a user secret
      operationId: createUserSecret
      security:
        - BearerJwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserSecretRequest"
      responses:
        "200":
          description: Secret created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSecretItem"
        "400":
          description: Name/value missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Project not owned by caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Secret name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/secrets/{id}:
    get:
      tags: [Secrets]
      summary: Get a single user secret
      operationId: getUserSecretById
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/SecretId"
      responses:
        "200":
          description: Secret item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSecretItem"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Secrets]
      summary: Update a user secret
      operationId: updateUserSecret
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/SecretId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserSecretRequest"
      responses:
        "200":
          description: Secret updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSecretItem"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Secret name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Secrets]
      summary: Delete a user secret
      operationId: deleteUserSecret
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/SecretId"
      responses:
        "200":
          description: Secret deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/projects/{id}/secrets:
    get:
      tags: [Projects]
      summary: List project-linked secrets (without secret values)
      operationId: listProjectSecrets
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project secrets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSecretItem"
        "403":
          description: No project access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Projects]
      summary: Create a project-scoped secret
      description: Owner-only endpoint.
      operationId: createProjectSecret
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSecretCreateRequest"
      responses:
        "200":
          description: Project secret created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSecretItem"
        "403":
          description: Owner only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/projects/{id}/secrets/{secretId}:
    put:
      tags: [Projects]
      summary: Update a project-scoped secret
      description: Owner-only endpoint.
      operationId: updateProjectSecret
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ProjectSecretId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectSecretUpdateRequest"
      responses:
        "200":
          description: Project secret updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSecretItem"
        "403":
          description: Owner only or secret not in this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project/secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Projects]
      summary: Delete a project-scoped secret
      description: Owner-only endpoint.
      operationId: deleteProjectSecret
      security:
        - BearerJwt: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/ProjectSecretId"
      responses:
        "200":
          description: Project secret deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Owner only or secret not in this project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project/secret not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Project ID
...
    SecretId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: User secret ID
    ProjectSecretId:
      name: secretId
      in: path
      required: true
      schema:
        type: string
      description: Project secret ID
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Standard human/agent login JWT
    BearerByoa:
      type: http
      scheme: bearer
      bearerFormat: BYOA
      description: Raw BYOA token with `byoa_` prefix
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PageInfo:
      type: object
      required: [limit, hasMore]
      properties:
        limit:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true
...
    SecretListItem:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
        projectName:
          type: string
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        lastUsedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SecretListResponse:
      type: object
      required: [items, totalCount, pageInfo]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SecretListItem"
        totalCount:
          type: integer
        pageInfo:
          $ref: "#/components/schemas/PageInfo"
    SecretListLegacyResponse:
      type: array
      items:
        $ref: "#/components/schemas/SecretListItem"
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    UserSecretItem:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        lastUsedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserSecretRequest:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        description:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
    UpdateUserSecretRequest:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        description:
          type: string
          nullable: true
        projectId:
          type: string
          nullable: true
...
    ProjectSecretItem:
      type: object
      required: [id, name, createdBy, permissions, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
          nullable: true
        name:
          type: string
        createdBy:
          type: string
        permissions:
          type: object
          additionalProperties: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        lastUsedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    ProjectSecretCreateRequest:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
        permissions:
          type: object
          additionalProperties: true
    ProjectSecretUpdateRequest:
      type: object
      properties:
        name:
          type: string
        value:
          type: string
        permissions:
          type: object
          additionalProperties: true
...